Rmoving hardcoded URLs in templates
The problem with this hardcoded, tightly-coupled approach is that it becomes challenging to change URLs on projects with a lot of templates. However, since you defined the name argument in the url() functions in the polls.urls module, you can remove a reliance on specific URL paths defined in your url configurations by using the {% url %} template tag.

Reported-by:Shanrui Yong
Signed-off-by: Shanrui Yong <shanrui.yong001@umb.edu>