Improving our view
We need to amend the get_queryset() method and change it so that it also checks the date by comparing it with timezone.now().
First we need to add an import:
from django.utils import timezone

and then we must amend the get_queryset method

Question.objects.filter(pub_date__lte=timezone.now()) returns a queryset containing Questions whose pub_date is less than or equal to - that is, earlier than or equal to - timezone.now.

Reported-by:Shanrui Yong
Signed-off-by: Shanrui Yong <shanrui.yong001@umb.edu>