Testing our new view
Add the "from django.core.urlresolvers import reverse" to polls/tests.py

and we’ll create a shortcut function to create questions as well as a new test class

First is a question shortcut function, create_question, to take some repetition out of the process of creating questions.

test_index_view_with_no_questions doesn’t create any questions, but checks the message: “No polls are available.” and verifies the latest_question_list is empty. Note that the django.test.TestCase class provides some additional assertion methods. In these examples, we use assertContains() and assertQuerysetEqual().

In test_index_view_with_a_past_question, we create a question and verify that it appears in the list.

In test_index_view_with_a_future_question, we create a question with a pub_date in the future. The database is reset for each test method, so the first question is no longer there, and so again the index shouldn’t have any questions in it.

Reported-by:Shanrui Yong
Signed-off-by: Shanrui Yong <shanrui.yong001@umb.edu>